Move interfaces to interfaces
